Add support for struct enums
Struct enums with a single field are not much different than tuple enums
with a single field. An example (niche) use case for this is, when you
want to bind a enum with `pyo3` but at the same time use `enum_dispatch`
with it. `pyo3` doesn't support tuple enums, while `enum_dispatch`
doesn't support struct enums.

With this, mixed enums (one-field struct, one-field tuple, unit) are
supported.